Add # as the mod selector symbol: mod#id.
This use to be ., but it created a bunch of annoying ambiguities with . for unary .selector functions. These could probably be resolved at check time? But why make it hard to understand the code without type checking? Instead, let's just introduce a symbol to use for selecting from a module.