Change block and block type grammar.
Blocks used to be {parm Type, parm Type | exprs}. This lead to ambiguity:
{x | y} could either be:
	- a block with a parameter x and an expression y, or
	- a block with an expression x | y.

In general, using most characters to separate the parameters from the expressions lead to some kind of ambiguity:

	- {|x| y}
		is it a parameter x with an expression y or an expression  ((|x) | y)?
	- {(x) (y)}
		is it a parameter x with an expression y or an expression ((x)(y))?
	- {x: y}
		is it a parameter x with an expresson y or a keyword call (x: y)?
		(note that {x : y} is OK with a space, but that's kind of subtle.)
	- {x. y}
		is it a parameter x with an expression y or a selector (x.y)?

Of course with ordered choice we can resolve these by defining which is preferred, but it's ugly to arbitrary choose one.

So, instead we move the parameters outside of the {}-braces. This is a bit more "noisy", because we need an opening delimiter, ( in this case, since we don't piggy-back on the opening {. However, it does not have the ambiguities. Further, it more-closely mimics what calls look like, with the parameters paranthesized just before the opening {.

There is no ambiguity with type definitions, becasue (...){...} was not a valid type before. Note, however, that we need to make sure to check for a FuncType before a paranthesized type in the ordered choice, and similarly in primary expressions for block literals. This hardly seems like a problem -- it's just 1 token of lookahead.