Add type aliases.
The Go folks make a strong argument for these for use in refactoring. Further, I would like to use type aliases as a way to make exported, opaque types. Instead of having to find a syntax to make struct fields or union cases un-exported, you can put them in an unexported type but give it an exported alias. Users of the module can reference the alias by name, but nothing else about the type, it's constructor, or built-in functions.