Greatly simplify struct and union lit checking, plus shorten common ident names.
Adds the concept of an "appropriate type" for a literal. Then split the type checking into two cases depending whether the expected type is appropriate to the literal or not.

Further, rename xyzLit to just lit and parserXyzLit to just parserLit, since these are very commonly used, and the longer names were hurting readability.