Remove named/unnamed type conversion; it's not quite right and we may simply not want it.
To be useful, this probably should probably convert between a named type and a fully literal type (no type names) if their fully literal forms are equivalent. But it would not be useful in the face of union types, which literals can only infer a single case of. So, maybe we should just skip this type of conversion and require an explicit type.