Add built-in selector functions.
This needed a few changes:
* Add built-in Selectors to the module scope
* Add Func.unifyRet and Func.unifyParm
* Add conversion between equal, literal types
* Change calls to always return a reference type wrapped in a deref node.

The last of these makes call expressions addressable. The reason this is needed is because selectors take a reference as their argument, but you should be able to call a selecton on the result of a function:

	type point [.x float64, .y float64]
	func makePoint() point
	var _ := makePoint().y

The only way to call .y on the point-type result of makePoint is if it can be converted to a reference.

This seems OK to me: the type checker assumes that essentially everything is in memory except an expression that converts something else to a reference. So you can add 1-level of reference, but not more than 1. Downstream code will make this same assumption anyway: everything is allocated in memory and only promoted to a register if it can be proven that its address is never needed