Change overload resolution to always unify returns after all parameters, never before.
The old logic was that if there is an expected type, the return would be unified with it before all parameters. This caused issues:

	var xs [int] := [1, 2, 3]
	var x &int := xs[1]

The call to xs[1] here, using the old logic, was an error. It would first unify the return for []([T])&T, binding T to &int (unification strips off all references before binding, so it can unify a hopefully convertible type). This binding propagated to the arguments, leading it to assume the call should be to []([&int], int)&&int, causing a conversion failure.

This change switches the logic to always unify the return type last, after the arguments. Now, in the code above, it sees that xs is [int], and the unified function correctly becomes []([int], int)&int.

This also seems to have more straight-forward behavior: the type is nearly always dictated by the first argument. The previous logic sometimes dictated it by the argument, but other times by the return.

Further, it simplifies the code: a bunch of unification code goes away for Selector, Switch, and Builtin.

In the future, we could look into doing something more sophistocated for unification with &. Instead of always stripping all leading &s, just strip the correct number of leading references from the type so that it matches the pattern. Otherwise, if the type has fewer than the pattern, fall back to stripping all and unifying like we do today.