Change unifyParam rules for references and fix literalType(Type) to exclude TypeVars.
The old rules unified the value types, assuming that reference conversion would then apply. This is broken trying to unify a type &T with a pattern U, as it would bind U=T. The problem is if the intension was U=&T, there was no way to achieve it.

The new rules are:
* if pat is a literal and type has a literal representation, unify type's literal with pat,
* else if pat is a literal and type has a literal representation, unify type with pat's literal,
* then, if pat has a greater reference depth than typ, unify typ with pat, with the mismatching number of references removed.

This means that if unifying type T with pattern &U, the unification will bind U=T, allowing for a possible, implicit reference conversion. Another example, unifying &[.foo int] with &&[.foo T] will bind T=int.