Built-ins work on defined types as well as basic and literal types.
Previously, built-in function only accepted basic types or literal types. So, +, -, *, /, and so forth only accepted basic types as their arguments, selectors and switches only took type literals as their arguments. Defined types were not allowed.

For literal types, this was mostly OK, because defined types can be automatically converted to the literal type. However, it meant that iface functions would not match selectors or switches no defined types, as these functions' arguments were always only type literals.

	func foo(_ T) : .length(T)int
	type my_array [int]
	func main() {
		x := my_array : [1, 2, 3],
		// This call was an error, since .length was .length([int])int,
		// which is not equal to .length(my_array)int.
		foo(x)
	}

For basic types, it meant that defined types could never use these operators, since there was no way to convert, and the types were not equal.

	type my_int int
	func main() {
		x := my_int : 4,
		y := my_int : 6,
		// This call was an error because +(int, int)int
		// is not +(my_int, my_int)my_int,
		// and there is not automatic conversion my_int -> int.
		x + y
	}

With this change, built-in functions exist for all relevant types, whether defined, basic, or literal.

More formally:

Arithmetic and comparison operators exist for all float and integer types.

Bitwise operators exist for all integer types.

Selector functions exist for all struct reference types.

Switch functions exist for all union reference types.

An integer type is a basic type int, int8, int16, int32, int64, uint, uint8, uint16, uint32, or uint64; or a defined type where the definition type is an integer type.

A float type is a basic type float32 or float64; or a defined type where the definition type is a float type.

A struct type is a struct literal type or a defined type where the definition type is a struct type. A struct reference type is a reference type referring to a struct type or a defined type where the definition type is a struct reference type.

A union type is a union literal type or a defined type where the definition type is a union type. A union reference type is a reference type referring to a union type or a defined type where the definition type is a union reference type.

Returning to the above examples:

	func foo(_ T) : .length(T)int
	type my_array [int]
	func main() {
		x := my_array : [1, 2, 3],
		// This call is OK, since .length(my_array)int exists.
		foo(x)
	}

and

	type my_int int
	func main() {
		x := my_int : 4,
		y := my_int : 6,
		// This call is ok: +(my_int, my_int)my_int.
		x + y
	}