Redesign and re-implement conversions.
Conversions (numeric and `[int8]`  to `string`) are no longer functions, but instead done using the type conversion grammar. For example, `int8(x)` becomes `int8 : x`.

In addition, implicit conversions of conversions expressions are disallowed. For example,

	func f(_ &int)
	...
	f(int : 8)

is now an error, since the implicit conversion of `int` to `&int` does not happen on the explicit conversion expression `int : 8`.