Add FrameType and Frame to flowgraph.
These are used to implement long returns. A long return must track to which frame it needs to return. The frames need some kind of identifier. These identifiers are of an opaque, small, non-empty type FrameType, and the identifier for the current frame can be retrieved as a Frame value.