In overload resolution, only check return convertibility if there is more than one candidate.
The check only checks for implicity type convertibility. However, if the parent of the call is an explicit conversion, then would fail in unintuitive cases. For example, when there is only one possible candidate, and its return can explicit convert.

Instead, this change makes overload resolution skip the return type conversion check entirely if there is only a single candidate and just picks the candidate.

Note that this approach could be taken further in the future, if needed. For example, if there are two candidates that fail to implicit convert to the desired type, but one would explicitly convert, then we could pick that one. But for now, we will go with this simpler approach.