Change the grammar for explicit type conversions.
The previous grammar used to use : for conversions. This was problematic because it required spaces on both sides of : to prevent ambiguities with keyword functions. Further, it didn't stand out much (in my opinion) when used within a keyword function call argument, because it somewhat blended in with the keywords themselves.

The new grammar makes two changes: 1) type conversions now use a double-colon, ::, and 2) the conversion expression has a lower precedence. Change #1 makes the conversion unambiguous even without spaces and makes it more distinguished within the argument of a keyword call. The second change makes it somewhat more simpler what the conversion is converting; it converts the entire expression to its right.