Make blockBuilder.Return have a Frame argument.
Previously the Frame was set on the Return after it was added to the BasicBlock. However, this broke used-by tracking. Used-by is set when adding an instruction to the BasicBlock, so setting Frame (a use) after the Return is added is too late to correctly update the used-by list of the Frame. By making it an argument, it more-closely matches all the other builder methods, and it correctly tracks the use.