Change switch case names to have the ? at the end instead of the beginning.
The original motivation for putting it at the beginning was to be similar to select identifiers which _begin_ with a period. But in practice it just looks strange. It looks much better at the end, becasue then it appears as if a question:

	x < y true? { do_true_thing() } false? { do_false_thing() }

Further, it better parallels keyword calls, which have their punctuation (the colon) at the end; switch calls are, at least syntactically, more similar to keyword calls.