Add an Obj struct to flowgraph/interp, instead of each value being its own object.
In the future we need to carry extra information with each object (namely, whether it has been deleted). This change rename Obj to Val and adds a new struct type Obj that contains a value. Now additional information about each object can be added to the Obj struct (a deletion bit) without needing to change every Val implementation.

While doing this, a few extra checks are added: panic if accessing an uninitialized value, and panic if setting the value of an object with a value of a different type. More such checks can get added in future changes.