Value.UsedBy and BasicBlock.In return copies to avoid changing during ranging.
There was a bug where value.users was changed while iterating across it in a range loop. Instead, UsedBy returns a copy of the slice so that we can iterate it while also modifying it. There was no similar known bug with BasicBlock.In, but change that one too for safety.