Fix inaccurate in/out func reference tracking.
Inlining had a bug in function reference tracking. When a function was inlined, the reference count to the inlined function was decremented. But if the function is called multiple times, then the first inlining of it would decrement the ref count to 0, which is wrong.

Instead, decrement reference counts during instruction removal in rmUnreach and rmDeletes.

Further, when inlining, the inlined function's reference counts should be added to the function into which it is inlined.