More thorough Loc setting for flowgraph instructions.
A bunch of instructions had loc.Loc{} for their location when they could instead be traced back to a correct source location.

This change actually will make the builder panic if trying to add a Call with no location, since the llvm backend assumes that calls have locations in many cases.