Add a temporary hack in pea_test.sh to work aroung adding main() to .a files.
The test code, in particular the main() function, should not live in the library .a files, but just the .main binary itself. The bug is that linking together .a files from multiple modules built in test mode will emit multiple main() functions and causing linker failures.

The temporary hack is to just remove all .a files and build everything from scratch in pea_test.sh. This is OK for now since there is very little code to build.