Fix a bug in block literal returns inferring a value from a reference type.
	x := &int :: 5
	b := (){x}

would give b the type (){int} instead of (){&int}. The reason is that the checker was unintensionally calling convert() on x, trying to convert it to typ=nil (which should not be valid). The result ended up being that all references (including references inside defined types) were stripped.

This showed up in the printer type, which is a defined reference type:

	type printer := &[.x, .y, .z] // roughly

Returning a printer from a block literal should return a reference to that single underlying printer, but it was unintensionally inserting a copy.