Allow explicit conversions between defined types and types with the same defined base type.
Defined base type (maybe the name of this will change to something better) is:
* if the type is a defined type, the defined base type of the type in its definition.
* if the type is a reference type, the reference to the defined base type of the referred type.
* otherwise, the type.

Essentially this allows:
	type t string
	string :: (t :: x)

and
	type t string
	t :: (string :: x)

and
	type t u
	type u string
	t :: (string :: x)

and
	type t &int
	t :: (&int :: x)

and so forth — conversions between defined types and the actual representation of that type. However, it does not look into struct fields or union cases. So for example these are not allowed:

	type t [.x u]
	type s [.x v]
	type u string
	type v string
	t :: (s :: x)	// *** error

and:
	type t [x? u]
	type s [x? v]
	type u string
	type v string
	t :: (s :: x)	// *** error