Add implicit conversion of any type to an empty struct.
This effectively just drops the value. It is part of the preparation for changing top-level expression checking. Currently top-level expressions have a nil want type. A follow-up change will make this the empty struct type. That will allow for top-level switch calls to have cases where each case block has a different type last expression — the inference will infer it to result in [.].