Move block result expression checking and conversion to checkBlockLit.
This was previously done in checkExprs. It was a bit of an awkward special case to handle in checkExprs, and it was a bit ridged because it didn't allow as much special handling of the block result expression.

A follow-up change will make checkExpr use [.] as the want type for all top-level expressions. This would prevent issues for block literals where sometimes we want the last expression to check with the result type, and some times (for inference) we want it to have a nil want type. By moving the result expression checking into checkBlockLit, it moves to a place where we more clearly know which case to use.