Change := to return the empty struct instead of the assigned value.
It's probably best not to encourage assignment as an expression, but more importantly, this helps a bit with switch expressions like:

	x true? {a := b} false? {x := y}

where previously this would not compile if the type of b and y did not match. As the result type of the functions would not match. Now the result type is the empty struct in both cases.