Fix panic on end-of-file loc endings.
The end of a loc is exclusive. If the location ends at the end of a file, its exclusive end loc will be the first byte of the next file. In this case, we were inappropriately panicing that it was a multi-file location.

This change adds a special case to avoid the panic, since this is not a bug.