Correctly handle recursive reference types.
Flowgraph neglected to correctly handle recursive reference types correctly. This change fixes that.

Since LLVM only supports recursive struct types, we implement recursive reference types by erasing the type on the recursive occurrence. So a recursive type `type x = &[/* refers to x somewhere in here */]` is converted to *struct{} on the recursive occurrence, as if `type x = &[/* instead *struct{} occurs in here]`.

In the flowgraph building, when accessing a struct field, if the field type is *struct{}, but the checker expected some type *T, then this must be a type-erased, recursive reference type.

We add a NumConvert to convert *struct{} to *T. LLVM and the flowgraph/interp now support NumConver to bitcast a pointer.

Further, in LLVM, the base value type of a Field instruction may be some `struct{/*...*/, i8*, /*...*/}`, but expected struct of the Field instruction is `struct{/*...*/, T*, /*...*/}`; this is _also_ an erased recursive reference type. Sometime during optimization the base value of the Field was swapped to a value that used the erased type, where the original field access did not have the erased type. We add a bitcast here too so that the LLVM types match.