Move conversions for type-erased struct fields from flowgraph to llvm.
Previously the conversion code was split: llvm handled conversion if the base type was unexpected, and flowgraph handled conversion of the field type was unexpected. In doing so, we increased the scope of NumConvert to also bitcast pointers, since flowgraph needed an op to communicate the conversion.

This change removes bitcasting from NumConvert. Instead of flowgraph commuticating the conversion, flowgraph always ensures that Field.BaseStruct has non-erased pointer fields, converted directly from the *checker.StructType. This gives llvm all the info it needs to handle both types of conversions.