Don't add the test main or main function to the module .a file.
Just add the exported non-main function defs to the .a file. They may be imported into another module. The test main and main functions are only added to their own .o files, linked just to make the binary for that individual module.

For main, but not test main, this doesn't matter; main should never be importet. But it seems consistent to just do the some for both.