Don't crash calling a function name for a function with an argument type error.
What was happening is a type that fails to check will contain nils. Call checking always assumes non-nil types, because it assumes scope lookups only return non-error definitions. But error definitions _were_ added to scopes. Now they are not. If there is an error checking a variable type or function signature, it is not added to the definition list of the module.