Move os to sys/fs and add comments to exported definitions.
All of the functions in os were file system specific, so just put them in a module about filesystems. It is sys/fs. This starts the convention that modules which wrap calls to the operating system are sub-directories of sys/.

Types not specific to the syste implementation and their supporting functions are in non-sys modules: io and fs.

Finally, io#write_result is re-worked to allow returning the number of bytes written even in the case of an error.