Move modules/algorithm/sort to modules/sort.
The word "algorithm" is pretty long. I was tempted to rename it "alg" to match "sys" (short for, of course, "system"). But I decided it was unclear what should actually go in "alg"/"algorithm". Hashing is an algorithm. Should hashing be in there? A bunch of the containers use algorithms. In some sense all programming is about an algorithm of some sort. So anyway, that directory name doesn't seem to make sense at the moment.