Add the sys/internal and sys/io modules; share common functions with sys/fs.
The sys/io module contains an opaque 'descriptor' type. The intent is that this type will remain completely opaque. I/O related sys# types will be able to return their opaque descriptor. And modules internal to sys/* can use sys/internal to read, write, and close descriptors.

Externally, descriptor will be used in places where any of the various system I/O types can be passed to the underlying operating system call. For example, a future sys#proc#command type will have stdin, stdout, and stderr fields that are the descriptor type. They will be internally used for a dup2 call to pipe stdin, stdout, and stderr for a new child process. This will allow files, pipes, network connections, and whatever types can be converted to a descriptor to be the stdXYZ files of a child process.

The sys/io module also contains a generic 'stream' type. This is a generic readable, writeable, closable type that also has a descriptor. It is used for stdin, stdout, stderr, and will be used for pipes. Files and network connections are very similar to streams, but use separate types, because they may have separate functions. For example, seek can work on files, but not streams or network connections.

The sys/internal module contains some functionality shared by sys/* modules. It is named 'internal' to indicate that it is not intended to be used outside of sys/* and would not be part of any future API guarantees.

Currently sys/internal has some error handling and file descriptor read, write, and close functions.