Fix incorrect TODO comment.
The TODO assumed that max_uint32 wasn't exported from basic; it was. The TODO bug should be fix, but the TODO is wrong. Instead, use the shared max_uint32 from basic, move the TODO into basic, and re-word it to say that max_uint32 should move to a shared int module.