Fix a bug, now allowing a ref type to sub for a type var creating a double-reference.
This code used to lead to a crash during type substitution when S=&string:

	func foo(s S) &S : [](S, int, int)S {
		return: &S :: s[5, 6]
	}
	func main() {
		str_ref := &string :: "",
		foo(str_ref)
	}

The reason is that &S, when S=&string, would be &&string. It is not normally possible to convert the return of s[5, 6], of type string, to &&string.

However, type variables are supposed to act like defined types. So the return of s[5, 6] should be type S=&string, which should be convertable to type &S=&&string.

To support this, we need a way to add a new reference, a new level of indirection. This is done by changing how called interface functions are implemented. Instead of substituting the call to be a call to the static interface argument function itself, we wrap the call in a new block. The block calls the static interface argument function, assigning its return to a local variable. Then this local variable is returned:

	func foo(s S) &S : [](S, int, int)S {
		return: &S :: {
			ret := s[5, 6], // ret introduces the extra &.
			ret, // &string return can convert to &&string.
		}()
	}
	func main() {
		str_ref := &string :: "",
		foo(str_ref)
	}

Somewhere in this mess, we end up needing the flowgraph to support &[.], though I don't quite recall why. So while at it, add support for &[.]. This may not be complete, but it is at least enough to cover the cases needed by this change.