Set recursive iface limit to 10.
A function could previously not satisfy its own interface. This meant that we could not allow things like printing nested arrays with:

	+=(printer, [T])printer : +=(printer, T)

Printing `[[1, 2, 3]]' would look find this and instantiate `+=(printer, [[int]])printer : +=(printer, [int])`. When trying to satisfy the interface, the same function was excluded from the search scope to prevent recursive, but this means the entire call would fail.

Now we allow the same function to be considered to satsify its own interface, but only up to a depth of 10. This means that up to 10-nested arrays, for example, can now be printed.