Remove exponential behavior resolving interface functions.
We previously broke loops in interface instantiation only when we hit a depth of 10 occurrences of the same function definition. A very generic function with interface arguments of its own name would create tight cycles that could lead to exponential behavior. The exponential expansion would be trimmed at a depth of 10, so it was not infinite recursion, but was extremely costly.

The example demonstrating this was:

	Func ++(t T, u U)printer : ++(printer, T)printer, ++(printer, U)printer
	Func ++(printer, int)printer
	Func ++(printer, float)printer

Every instantiation of ++, I believe, was expanding the full 2^10 binary tree of ++(T,U) functions with its two interface arguments, even if they would trivially never match.

This change breaks cycles where a function *instance* tries to instanciate itself as its interface argument. This can only be an impossible cycle.