Refactor printing to use print and ++ instead of +=.
Previously printable types implemented +=(printer, T)printer, and then they could be added to a printer with +=.

With this change, printable types implement print(printer,T)printer. The print module makes a ++(printer,T)printer function for anything implementing print(printer,T)printer, and a ++(T,U)printer for any pair of functions <T, U> that implement ++(printer,T)printer and ++(printer,U)printer respectively.

For converting printable types to a string, instead of using a finite set of overloads of `print#string(T, U, V, …)string : +=(printer, T)printer, ...`, print#string just takes a printer. You can create a printer by chaining ++, so:
	print#string(a, b, c)
becomes:
	print#string(a ++ b ++ c)
but instead of being limited to 5 arguments, it can be chained to infinity:
	print#string(a ++ b ++ c ++ d ++ e ++ f ++ g ++ h ++ i ++ …)

A downside to this approach, however, is that to get ++(T,U)printer, the caller most likely needs to capital Import the print module. This is particularly annoying, because the print module contains some functions that have names likely to collide with identifiers in the including modules: bytes, for example. This will need to be solved.