Fix nondetermministic flowgraph output.
The function topological sort iterated over map entires. It did this by first putting them in a slice and sorting it by FuncDef.Name to fix map order non-determinism. But FuncDef.Name is not unique, so it didn't actually fix the ordering problem. This change sorts by SourceName which is unique, fixing the issue.