Cache test results in pea_test.sh.
Currently pea_test.sh is quite slow. Much of the issue is the regexp exhaustive tests, which take some time to run. This change tries to address that by caching test results so that tests don't need to be rerun if the corresponding test binary hasn't changed.

It works by storing test results in a cache file named by the sha1sum of the test binary name. If the cache file exists, skip running the test and just dump the output instead. If the file does not exist, run the test, outputting to the file.