Deduplicate panic file string constants in .ll files.
Before this change, each occurrence of panic emitted a string constant with its source file name. Many panics can appear in the same file, and in some cases this was leading to 10s of kilobytes of .ll file being devoted to just the panic file strings.

In practice, most of these were duplicates, so deduplicate them.