Define flowgraph and llvm types for structs of pointer-to-struct types.
Previously, if a defined type was a reference type, it did not get any defined type in the LLVM output. The literal representation of these types in LLVM .ll files could be quite verbose.

This change makes a new type def for unnamed structs in defined pointer-to-struct types. Now in the .ll output, the type is not named by a pointer to the struct literal, but a pointer to the struct's named type.

For the current container/hash_map module, this reduces the .ll file by nearly 100KB (of 850KB originally). It also improves readability of the .ll files.