Don't compute flowgraph for some other-module functions.
The goal is to allow functions that may be beneficial for inlining, but to omit other functions so that we don't waste time re-compiling them.

Currently, the logic is to allow functions that have any function-type argument (these could be passed block literals, and thus benefit from inlining) or fewer than 10 "statements" (a rough approximation for the size of the function).