Change inlining condition to allow inlining calls with block literal arguments.
The previous test was complex and poorly documented, and it was too strict, preventing inlining of for-loops, for example (they called if:then:, so weren't considered a leaf).

Now the test for inlinability is better documented and traced, and it allows inlining all calls that have block literal arguments. This lets for-loops, for example, to be inlined even though they aren't leaf calls.

There should probably be more tweaking here, but for now it's quite the improvement. The regexp tests on my laptop go from ~12s down to ~0.8s.