Improve conservative escape analysis for struct, union, and array types.
Previously if the type was used by a Field, Case, or Index that was not Load-only, it would be considered to escape. The point was to consider structs, unions, and arrays as escaping if a field, case, or index escaped. But Load-only is too strong of a condition. This change allows both Loads and Stores-to.

Let's consider just structs, for example. Every allocated struct always has its fields initialized (with a Store). Storing a value to the struct field certainly does not make the field or struct base escape.