Don't mark return values as escaping.
Technically the address can escape if it is passed in a capture block of an escaping function. But we always ensure that such a function checks that the parent frame is a parent on the current stack before writing to the return location (otherwise it panics). So we can never write to such an escaped return location, and thus it is safe to move it to the stack.

This reduces the number of heap allocations for the current regexp module from 1083 to 446.