Fix a bug in regexp ambiguous subexpr matching.
The match was a correct match, but the package should match the same subexpressions as a backtracking algorithm, but it was coming up with different matches.

The issue was that match? was tested in add_thread: not step:. In the face of a fork, if the LHS of the fork matched, it would truncate the cur queue, and then the recursion would unwind and add the RHS of the fork. But this RHS should have been truncated, since the LHS has priority.

The fix was to move match? into step: so it happens at the next iteration of the VM. This way, when it truncates the queue, it will correctly truncate the thread(s) added by the RHS.