Improve conservative escape analysis for structs, unions, and arrays with composite fields.
This change is no that conservative escape analysis doesn't mark a struct, union, or array as escaping if one of its fields, cases, or indices is copied to. Change 18a91e3 did this for store-tos, but for composite fields, the initialization of a field will be a copy-to, not a store-to.

This reduces the current runtime of the regexp tests on my laptop from ~18.9s to ~15.5s.