For a module-tagged call, add the tagged module's exported definitions to the scope.
This can help to reduce the number of Capital Imports by implicitly adding those modules' definitions to the scope in cases where it seems likely that they are expected.

For example, when calling `print#string(…)`, the `print` module's exported definitions are in scope while checking `…`. This allows to lower-case import the "print" module instead of Capital Importing it to get access to `func ++(T, U)printer`.