Add finalizer registration to the pea runtime.
Common caveats of finalizers apply here. There is little guarantee on when the finalizer runs besides that it is after the object becomes unreachable. The intent is for compatibility with libraries that require manual memory management. For example pthread_mutex_destroy, which must be called before the mutex memory is reclaimed or else it will cause memory leaks.