Make late escape analysis recur for Field, Case, and Index bases.
When doing late escape analysis for a struct , union, or array base value, if its Field, Case, or Index users were only initialized (roughly), then we didn't consider those users to escape the base. But we can do something better, which is recursively compute escaping of the Field, Case, or Index instead. This allows more values to be moved to the stack.

Among other things, this change makes it so that optionals aren't heap-allocated when their data field is accessed by a copy.