Use ring_buffer instead of vector for regexp run queues.
These are logically queues (push to the back, pop from the front). We were using vector here, implementing the "pops" by instead just iterating the current run queue and after iterating it entirely, we'd clear: it.

The problem with that approach is that threads (really just their mem) completed on the current step would be both in the run queue, in the portion already iterated over, _and_ on the next_run queue for the next step. When a match is hit, it cleared the current run queue. None of the memory for the cleared threads were added to the free_mem list, and it couldn't be easily added, becasue we didn't know where the matching thread was in the run queue, and couldn't add the entire run-queue to free_mem, since some was not free, but in next_run.

Anyway, this change has the invariant that, after step, any thread is only in a single queue: run or next_run, never both. This means that when we hit a match and want to clear run, we can add the mem of each thread remaining in run to free_mem.

The performance change here is pretty minimal. Actually, it may be slightly negative (hard to tell, since there is some noise in the test runtimes), because ring_buffer is a bit more costly than vector. But! this will make it it easier in a folllow-up change to share this extra memory on cached VMs, possibly greatly reducing allocations in the face of multiple calls to exec.