Don't initialize the GC before calling fork() in tests.
The GC uses pthreads. If we use any of the GC functions, it can spawn threads. GC threads are incompatible with fork(), so if you invoke the GC before calling fork(), you cannot safely use the GC in the forked child. This change moves Pea module initialization into the child-process-run tests functions and makes it so that the test main() only calls the test runner, and does not allocate or otherwise initialize the GC.

This lets us remove HANDLE_FORK support from the GC. HANDLE_FORK was a best-effort attempt to fix the problem with fork(), but the comments around it seem to indicate that it wasn't a surefire fix. Further, HANDLE_FORK support disabled incremental GC.