Rename mod.Loader to mod.Root and Loader.Load to Root.Get.
These names are more clear. The package just reads the module information from the module directory hierarchy. It doesn't do any compilation or linking or anything. The former names may have made that confusing.

This change also adds Root as a field of Mod so that peac doesn't need to pass an extra Root parameter around.