Track flowgraph.Type.SourceName, a unique identifier for checker.DefTypes.
Multiple different checker types can convert to the same flowgraph type. Currently, bool is an int64 when converted to flowgraph (eventually it will be uint8, but currently it's not). If a there is a DefType instantiated once with bool and once with int64, then it will have two TypeInsts that both convert to the exact same flowgraph.StructType.

This isn't a problem until LLVM output is created. Both of these defined StructTypes will generate the same LLVM type definition.

There are two options to fix this:
1) LLVM checks for duplicate type name strings before defining them, and eliminates duplicates, or
2) Find a different name to use that is unique for each type.

This CL implements Option 2 by tracking the checker name of each defined struct type in flowgraph.StructType.SourceName. LLVM can use these unique SourceNames instead of the non-unique .String() output.

This seems preferable to Option 1, because it maintiains more information late into the compilation (namely that they are different source types).